App/requirements.txt
fastapi
uvicorn
celery[rabbitmq]
pika
kombu
python-environ
requests
pydantic